treewide: Import minimal Xiaomi kernel changes for MI 9SE
The Patch based on QualComm release TAG: LA.UM.8.8.r1-07800-SDM710.0 .
The kernel config file used is grus_user_defconfig.

+++ arch/x86/kvm/vmx.c
@@ -8207,7 +8207,7 @@ static bool nested_vmx_exit_handled(struct kvm_vcpu *vcpu)
 		return true;
 	}
 
-	switch ((u16)exit_reason) {
+	switch (exit_reason) {
 	case EXIT_REASON_EXCEPTION_NMI:
 		if (is_nmi(intr_info))
 			return false;
